fixed rocket positioning
